#3: Move buildProductVariantAssetsUpdateActions back to ProductVariantUpdateActionUtils.
