+++ src/main/java/com/commercetools/sync/products/utils/ProductVariantAssetUpdateActionUtils.java
@@ -8,7 +8,6 @@ import io.sphere.sdk.models.Asset;
 import io.sphere.sdk.models.AssetDraft;
 import io.sphere.sdk.products.Product;
 import io.sphere.sdk.products.ProductVariant;
-import io.sphere.sdk.products.ProductVariantDraft;
 import io.sphere.sdk.products.commands.updateactions.AddAsset;
 import io.sphere.sdk.products.commands.updateactions.ChangeAssetName;
 import io.sphere.sdk.products.commands.updateactions.ChangeAssetOrder;
@@ -24,7 +23,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import static com.commercetools.sync.commons.utils.CommonTypeUpdateActionUtils.buildUpdateAction;
@@ -37,35 +35,10 @@ import static java.util.stream.Collectors.toMap;
 //TODO: CONSIDER CHANGE NAME.
 // TODO: Add tests.
 public final class ProductVariantAssetUpdateActionUtils {
-    /**
-     * Compares the {@link List} of {@link AssetDraft}s of a {@link ProductVariantDraft} and a
-     * {@link ProductVariant} and returns a {@link List} of {@link UpdateAction}&lt;{@link Product}&gt;. If both the
-     * {@link ProductVariantDraft} and the {@link ProductVariant} have identical list of assets, then no update action
-     * is needed and hence an empty {@link List} is returned.
-     *
-     * @param oldProductVariant the {@link ProductVariant} which should be updated.
-     * @param newProductVariant the {@link ProductVariantDraft} where we get the new list of prices.
-     * @return a list that contains all the update actions needed, otherwise an empty list if no update actions are
-     * needed.
-     */
-    @Nonnull
-    public static List<UpdateAction<Product>> buildProductVariantAssetsUpdateActions(
-        @Nonnull final ProductVariant oldProductVariant,
-        @Nonnull final ProductVariantDraft newProductVariant,
-        @Nonnull final ProductSyncOptions syncOptions) {
 
-        return ofNullable(newProductVariant.getAssets())
-            .map(newAssetDrafts ->
-                buildProductVariantAssetsUpdateActions(oldProductVariant, newAssetDrafts, syncOptions))
-            .orElseGet(() ->
-                oldProductVariant.getAssets().stream()
-                                 .map(oldAsset ->
-                                     RemoveAsset.ofVariantIdWithKey(oldProductVariant.getId(), oldAsset.getKey(), true))
-                                 .collect(Collectors.toList()));
-    }
 
     @Nonnull
-    private static List<UpdateAction<Product>> buildProductVariantAssetsUpdateActions(
+    static List<UpdateAction<Product>> buildProductVariantAssetsUpdateActions(
         @Nonnull final ProductVariant oldProductVariant,
         @Nonnull final List<AssetDraft> newProductVariantAssetDrafts,
         @Nonnull final ProductSyncOptions syncOptions) {
